+++ src/error.rs
@@ -61,6 +61,8 @@ pub enum FormatErr {
     InvalidFormula(String),
     #[error("invalid boolean expression: {0}")]
     InvalidBooleanExpression(String),
+    #[error("invalid attribute: {0}")]
+    InvalidAttribute(String),
     #[error("encrypted data size cannot be less than {0} bytes")]
     InvalidEncryptedDataSize(String),
     #[error("invalid encrypted data")]